Establish comprehensive test infrastructure with 140+ validated tests across all CFD dimensions (#80)
* Initial plan

* Remove documentation clutter and fix core compilation errors


* Fix all compilation errors across the workspace


* Convert magic numbers to named constants and fix naming conventions


* Fix critical Clippy errors and improve code quality systematically


* Achieve zero warnings in core crate and implement proper GPU resource management


* Fix all critical test compilation errors and achieve comprehensive test coverage


* Add comprehensive cfd-3d tests and fix failing cfd-core test - all crates now have validated working tests


* Expand cfd-1d test suite with comprehensive millifluidics tests


---------

+++ crates/cfd-core/src/gpu/validation_tests.rs
@@ -69,7 +69,7 @@ mod tests {
             .collect();
 
         let mut gpu_result = vec![0.0; field.len()];
-        gpu_ops.laplacian_2d(&field, nx, ny, dx, dy, &mut gpu_result);
+        gpu_ops.laplacian_2d(&field, nx as usize, ny as usize, dx, dy, &mut gpu_result);
 
         // Interior points should have Laplacian ≈ 4
         // Boundary handling may differ, so only check interior
